requirements.txt
Pillow>=9.0.0,<10.0.0
python-magic>=0.4.0,<0.5.0